changed Min_width type to string
According to the i3bar json protocol spec, the min_width field can be either a
string or an int. It's not easy in Go to parse a field that can have two
different types, so I chose to only support a single type for the min_width
field: the string type. The reason the string type was chosen over the int type
is that the string type for min_width (which specifies character-width) seems
more generally useful than the int type (which specifies pixel-width).

One wonders why the i3bar protocol creator(s) though it was a good idea to allow
multiple types for the same field.